+++ implementations/c/ziggurat.h
@@ -1,9 +1,9 @@
-float r4_exp ( unsigned long int *jsr, int ke[256], float fe[256], 
+float r4_exp ( unsigned int *jsr, int ke[256], float fe[256], 
   float we[256] );
 void r4_exp_setup ( int ke[256], float fe[256], float we[256] );
-float r4_nor ( unsigned long int *jsr, int kn[128], float fn[128], 
+float r4_nor ( unsigned int *jsr, int kn[128], float fn[128], 
   float wn[128] );
 void r4_nor_setup ( int kn[128], float fn[128], float wn[128] );
-float r4_uni ( unsigned long int *jsr );
-unsigned long int shr3 ( unsigned long int *jsr );
+float r4_uni ( unsigned int *jsr );
+unsigned int shr3 ( unsigned  int *jsr );
 void timestamp ( void );
